implementing field mapping input
